// src/services/controlPlaneApi/gen/models/gateway.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project API
 * Project description API
 *
 * The version of the OpenAPI document: undefined
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Authentication } from './authentication';
// May contain unused imports in some cases
// @ts-ignore
import { ConsumeOptions } from './consume-options';
// May contain unused imports in some cases
// @ts-ignore
import { ProduceOptions } from './produce-options';

/**
 * 
 * @export
 * @interface Gateway
 */
export interface Gateway {
    /**
     * 
     * @type {string}
     * @memberof Gateway
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Gateway
     */
    'type'?: GatewayTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof Gateway
     */
    'topic'?: string;
    /**
     * 
     * @type {Authentication}
     * @memberof Gateway
     */
    'authentication'?: Authentication;
    /**
     * 
     * @type {Array<string>}
     * @memberof Gateway
     */
    'parameters'?: Array<string>;
    /**
     * 
     * @type {ProduceOptions}
     * @memberof Gateway
     */
    'produceOptions'?: ProduceOptions;
    /**
     * 
     * @type {ConsumeOptions}
     * @memberof Gateway
     */
    'consumeOptions'?: ConsumeOptions;
}

export const GatewayTypeEnum = {
    produce: 'produce',
    consume: 'consume'
} as const;

export type GatewayTypeEnum = typeof GatewayTypeEnum[keyof typeof GatewayTypeEnum];